refactor(tests): split test_cleanup.py into a focused cleanup package (#443)
The 1122-line tests/teatree_core/test_cleanup.py mixed four separable
concerns in one file, which made it slow to open and hard to navigate.
Split it verbatim into a tests/teatree_core/cleanup/ package, one file
per concern, each under the 500-LOC module-health ceiling:

- test_cleanup_worktree.py — cleanup_worktree orchestration with the
  git layer mocked (TestCleanupWorktree)
- test_classify_branch_commits.py — classify_branch_commits bucketing
- test_real_git_integration.py — real git worktree removal (#460)
- test_recovery.py — #835 dirty/unpushed recovery capture
- _shared.py — the GIT_*-stripped runner the integration and recovery
  files both use, lifted unchanged

Pure relocation: no production change, no assertion or body change
(AST-identical), only import-path adjustments and verbatim helper
lifting. Same 41 tests collected, order-stable.

Relates-to #443
